Fix resoloution of aliased parent aliases.
Fixes https://github.com/pulumi/pulumi/issues/13324.
