+++ changelog/pending/20230713--engine--fix-alias-resoloution-when-parents-where-also-aliased.yaml
@@ -1,0 +1,4 @@
+changes:
+- type: fix
+  scope: engine
+  description: Fix alias resoloution when parent alieses where also aliased.
